alg-yescrypt-opt.c: Cast output of _mm_cvtsi128_si32 to uint32_t.
The non-__SSE2__ version of this code already has the same cast.
I'm not sure why this broke in CI, when it didn't on my workstation,
but I already had this fix in an experimental branch for other reasons.

+++ lib/alg-yescrypt-opt.c
@@ -252,7 +252,7 @@ static inline void salsa20_simd_unshuffle(const salsa20_blk_t *Bin,
 	XOR_X(in) \
 	SALSA20_wrapper(out, SALSA20_8ROUNDS)
 
-#define INTEGERIFY _mm_cvtsi128_si32(X0)
+#define INTEGERIFY (uint32_t)_mm_cvtsi128_si32(X0)
 
 #else /* !defined(__SSE2__) */
